Refactor Arithmetic interface and add log and isPrime methods
Refactored the Arithmetic interface in the Library module to include two new methods: log and isPrime. The log method returns the ceiling of the base 2 logarithm of a given number, while the isPrime method checks if a number is prime. These changes were made to enhance the functionality of the Arithmetic interface.